Corrección de errores y bugs
Corrección de errores y bugs
