+++ weblaruex/timetrackpro/templates/insertar-registro-diario.html
@@ -42,8 +42,6 @@
 
                   <input class="form-control" type="datetime-local" id="hora" name="hora" value="{{registro.hora|date:'Y-m-d H:i' }}" required>
                   </div>
-                  
-
 
                 </div>
                 <div class="row">
